refactor: extract embedded server module
